# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReleasePilot API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analyses": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Recent",
        "operationId": "recent_api_v1_analyses_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisSummary"
                  },
                  "title": "Response Recent Api V1 Analyses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Start",
        "operationId": "start_api_v1_analyses_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Start Api V1 Analyses Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/github/callback": {
      "get": {
        "tags": [
          "github"
        ],
        "summary": "Callback",
        "operationId": "callback_api_v1_github_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 300,
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/connect": {
      "post": {
        "tags": [
          "github"
        ],
        "summary": "Connect",
        "operationId": "connect_api_v1_github_connect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Connect Api V1 Github Connect Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analyses/{analysis_id}/status": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Status",
        "operationId": "status_api_v1_analyses__analysis_id__status_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Status Api V1 Analyses  Analysis Id  Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/retry": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Retry",
        "operationId": "retry_api_v1_analyses__analysis_id__retry_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Retry Api V1 Analyses  Analysis Id  Retry Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisSummary": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "source_mode": {
            "type": "string",
            "enum": [
              "demo",
              "github"
            ],
            "title": "Source Mode"
          },
          "recommendation": {
            "type": "string",
            "enum": [
              "GO",
              "CAUTION",
              "NO_GO"
            ],
            "title": "Recommendation"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "confidence_score": {
            "type": "integer",
            "title": "Confidence Score"
          },
          "head_sha": {
            "type": "string",
            "title": "Head Sha"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "state",
          "id",
          "service_name",
          "scenario",
          "source_mode",
          "recommendation",
          "risk_score",
          "confidence_score",
          "head_sha",
          "created_at"
        ],
        "title": "AnalysisSummary"
      },
...
      "ConnectionInput": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "pattern": "^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$",
            "title": "Full Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "public",
              "app"
            ],
            "title": "Mode",
            "default": "app"
          },
          "installation_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Installation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "full_name"
        ],
        "title": "ConnectionInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StartInput": {
        "properties": {
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "github"
            ],
            "title": "Mode",
            "default": "demo"
          },
          "scenario": {
            "type": "string",
            "enum": [
              "safe",
              "failed_ci",
              "missing_evidence"
            ],
            "title": "Scenario",
            "default": "failed_ci"
          },
          "base_ref": {
            "type": "string",
            "maxLength": 200,
            "title": "Base Ref",
            "default": ""
          },
          "head_ref": {
            "type": "string",
            "maxLength": 200,
            "title": "Head Ref",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "service_id"
        ],
        "title": "StartInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}